Мотовилова ПрИ-201 Модуль 2

+++ task03/src/com/example/task03/Task03.java
@@ -2,7 +2,7 @@ package com.example.task03;
 
 public class Task03 {
 
-    public static int getMetreFromCentimetre(int centimetre) {
+    public static double getMetreFromCentimetre(double centimetre) {
 
         return centimetre/100;
     }
@@ -11,7 +11,7 @@ public class Task03 {
         // Здесь вы можете вручную протестировать ваше решение, вызывая реализуемый метод и смотря результат
         // например вот так:
 
-        int result = getMetreFromCentimetre(345);
+        double result = getMetreFromCentimetre(345.00);
         System.out.println(result);
 
     }
